refactor(core): move `Compilation::update_module_graph` to its own file
Part of #5856
